+++ terraform_compliance/common/defaults.py
@@ -20,6 +20,7 @@ class Defaults(object):
         self.no_failure_tags = ['warning', 'no-failure', 'no-fail', 'info', 'nofailure', 'nofail', 'dontfail']
         self.case_sensitive_tags = ['case-sensitive', 'case_sensitive', 'casesensitive', 'case-sensitivity']
         self.no_skip_tags = ['noskip', 'no-skip', 'dontskip', 'failonskip', 'fail-on-skip']
+        self.exclude_resources_tags = ['exclude', 'ignore']
 
         if '--no-ansi' in sys.argv or not sys.stdout.isatty():
             self.skip_colour = \